refactor(subsurface): extract generic STAC symbology legend to shared module for route reuse
